requirements.txt
mediapipe
opencv-python
numpy
matplotlib
pygame
pyserial
librosa
pydub
torch
torchaudio
facial-emotion-recognition
python-vlc
streamlit